+++ opbeat/utils/traces.py
@@ -4,6 +4,11 @@ import time
 from datetime import datetime
 from opbeat.utils.encoding import force_text
 
+from collections import defaultdict
+import threading
+import time
+from datetime import datetime
+
 
 class Trace(object):
     def _decode(self, param):
@@ -31,24 +36,57 @@ class Trace(object):
         return (self.transaction, self.parents, self.signature, self.kind)
 
 
-class TracesStore(object):
-    thread_local = threading.local()
+class _RequestList(object):
+    def __init__(self, transaction, response_code, minute):
+        self.transaction = transaction
+        self.response_code = response_code
+        self.minute = minute
+        self.durations = []
 
+    @property
+    def fingerprint(self):
+        return (self.transaction, self.response_code, self.minute)
+
+    def add(self, elapsed):
+        self.durations.append(elapsed)
+
+    def as_dict(self):
+        return {
+            "transaction": self.transaction,
+            "result": self.response_code,
+            "timestamp": datetime.utcfromtimestamp(self.minute).isoformat() + "Z",
+            "durations": self.durations
+        }
+
+
+class RequestsStore(object):
     def __init__(self, collect_frequency):
         self.cond = threading.Condition()
+        self.thread_local = threading.local()
+        self.items = {}
         self._traces = {}
         self.collect_frequency = collect_frequency
         self._last_collect = time.time()
 
+    def add(self, elapsed, transaction, response_code):
+        with self.cond:
+            requestlist = _RequestList(transaction, response_code,
+                                       int(time.time()/60)*60)
+
+            if requestlist.fingerprint not in self.items:
+                self.items[requestlist.fingerprint] = requestlist
+
+            self.items[requestlist.fingerprint].add(elapsed)
+            self.cond.notify()
 
     def get_all(self, blocking=False):
         with self.cond:
             # If blocking is true, always return at least 1 item
             while blocking and len(self._traces) == 0:
                 self.cond.wait()
-            traces, self._traces = self._traces, {}
+            items, self.items = self.items, {}
         self._last_collect = time.time()
-        return traces.values()
+        return [v.as_dict() for v in items.values()]
 
     def should_collect(self):
         return (
@@ -58,7 +96,7 @@ class TracesStore(object):
 
     def __len__(self):
         with self.cond:
-            return len(self._traces)
+            return sum([len(v.durations) for v in self.items.values()])
 
     def request_start(self):
         self.thread_local.transaction_start = time.time()
@@ -68,15 +106,12 @@ class TracesStore(object):
     def set_transaction_name(self, transaction_name):
         self.thread_local.transaction_name = transaction_name
 
-    def set_response_code(self, code):
-        self.thread_local.response_code = code
-
-    def request_end(self):
+    def request_end(self, response_code):
         if hasattr(self.thread_local, "transaction_start"):
             start = self.thread_local.transaction_start
             elapsed = (time.time() - start)*1000
 
-            self.add(elapsed, 0, "request", "transaction.web", [], None)
+            self.add(elapsed, self.thread_local.transaction_name, response_code)
 
     @contextlib.contextmanager
     def trace(self, signature, kind, collateral):
@@ -98,11 +133,11 @@ class TracesStore(object):
         parents = [s[1] for s in sig_stack]
         elapsed = (time.time() - abs_start_time)*1000
 
-        self.add(elapsed, rel_start_time, signature, kind,
-                               ("request", ) + tuple(parents), collateral)
+        self.add_trace(elapsed, rel_start_time, signature, kind,
+                       ("request", ) + tuple(parents), collateral)
 
-    def add(self, duration, relative_start, signature, kind, parents,
-            collateral):
+    def add_trace(self, duration, relative_start, signature, kind, parents,
+                  collateral):
         transaction = getattr(self.thread_local, 'transaction_name', None)
         print signature[:20], relative_start, duration
         trace = Trace([duration], [relative_start], transaction, signature,