+++ CHANGES.md
@@ -1,3 +1,12 @@
+## 2.0.1 ##
+ * Fixed a bug when `__name__` does not exist for view_func given to OpbeatAPMMiddleware
+
+## 2.0 ##
+ * Async support was refactored, removed old AsyncClient classes
+ * OpbeatAPMMiddleware was added to time requests
+ * `APP_ID` now configurable from environment
+ * Lots of cleanups 
+
 ## 1.4 ##
  * added preliminary support for Django 1.8
  * dropped support for Django 1.3